additional config options

+++ src/post_build.rs
@@ -34,7 +34,7 @@ use std::error::Error;
 #[cfg(target_os = "macos")]
 use std::fs::{create_dir_all, remove_dir_all};
 use std::fs::{rename, File};
-use std::path::Path;
+use std::path::{Path, PathBuf};
 use std::process;
 
 use crate::config::{read_config, BuildError, Config};
@@ -47,43 +47,66 @@ use crate::config::{read_config, BuildError, Config};
 /// 5. Moves plug-in to FileMaker Extensions folder.
 /// 6. Launches FileMaker.
 pub fn bundle_plugin() -> Result<(), Box<dyn Error>> {
-    let config_path =
-        Path::new(option_env!("CRATE_MANIFEST_DIR").ok_or("CRATE_MANIFEST_DIR not set")?);
-    let config = read_config(config_path)?;
-    clear_log_file()?;
+    let config = read_config()?;
+    clear_log_file(&config)?;
     bundle_plugin_command(&config)?;
     launch_filemaker(&config)?;
     Ok(())
 }
 
-fn clear_log_file() -> Result<(), Box<dyn Error>> {
-    let user_dirs = UserDirs::new().ok_or(BuildError::LogFile)?;
-    let dir = user_dirs.desktop_dir().ok_or(BuildError::LogFile)?;
-    let path = Path::join(&dir, "plugin.log");
+fn clear_log_file(config: &Config) -> Result<(), Box<dyn Error>> {
+    if !config.log.clear_on_launch {
+        return Ok(());
+    }
+    let path: PathBuf;
+    if config.log.path.is_none() {
+        let user_dirs = UserDirs::new().ok_or(BuildError::LogFile)?;
+        let dir = user_dirs.desktop_dir().ok_or(BuildError::LogFile)?;
+        path = Path::join(&dir, "plugin.log");
+    } else {
+        path = Path::new(&config.log.path.as_ref().unwrap()).to_path_buf();
+    }
     let _ = File::create(path)?;
     Ok(())
 }
 
 #[cfg(target_os = "macos")]
 fn launch_filemaker(config: &Config) -> Result<(), Box<dyn Error>> {
-    process::Command::new("open")
-        .arg(&config.filemaker.bin_path)
-        .spawn()?;
+    if config.filemaker.bin_path.is_some() && config.filemaker.launch {
+        process::Command::new("open")
+            .arg(&config.filemaker.bin_path)
+            .spawn()?;
+    }
     Ok(())
 }
 
 #[cfg(target_os = "windows")]
 fn launch_filemaker(config: &Config) -> Result<(), Box<dyn Error>> {
-    process::Command::new(&config.filemaker.bin_path).spawn()?;
+    if config.filemaker.bin_path.is_some() && config.filemaker.launch {
+        process::Command::new(config.filemaker.bin_path.as_ref().unwrap()).spawn()?;
+    }
     Ok(())
 }
 
 #[cfg(target_os = "windows")]
 fn bundle_plugin_command(config: &Config) -> Result<(), Box<dyn Error>> {
+    if !config.plugin.bundle {
+        return Ok(());
+    }
     let out_dir = option_env!("CRATE_OUT_DIR").ok_or("CRATE_OUT_DIR not set")?;
-    let package_name = get_package_name()?;
-    let from = format!("{}\\{}.dll", out_dir, package_name);
-    let to = format!("{}/{}.fmx64", config.filemaker.ext_path, config.plugin.name);
+    let mut package_name = get_package_name()?;
+    package_name.push_str(".dll");
+    let from = Path::new(out_dir).join(package_name);
+
+    let mut plugin_name = config.plugin.name.to_owned();
+    plugin_name.push_str(".fmx64");
+
+    let to: PathBuf;
+    if config.filemaker.ext_path.is_some() && config.plugin.move_to_ext {
+        to = Path::new(config.filemaker.ext_path.as_ref().unwrap()).join(plugin_name);
+    } else {
+        to = Path::new(&out_dir).join(plugin_name);
+    }
     rename(from, to)?;
 
     Ok(())
@@ -91,21 +114,33 @@ fn bundle_plugin_command(config: &Config) -> Result<(), Box<dyn Error>> {
 
 #[cfg(target_os = "macos")]
 fn bundle_plugin_command(config: &Config) -> Result<(), Box<dyn Error>> {
-    let out_dir = env!("CRATE_OUT_DIR");
+    if !config.plugin.bundle {
+        return Ok(());
+    }
+    let out_dir = option_env!("CRATE_OUT_DIR").ok_or("CRATE_OUT_DIR not set")?;
     let package_name = get_package_name()?;
-    let plugin_path = format!(
-        "{}/{}.fmplugin",
-        config.filemaker.ext_path, config.plugin.name
-    );
+    let mut plugin_name = config.plugin.name.to_owned();
+    plugin_name.push_str(".fmplugin");
+    let plugin_path: PathBuf;
 
-    remove_dir_all(&plugin_path).ok();
-    create_dir_all(&plugin_path)?;
+    if config.plugin.move_to_ext && config.filemaker.ext_path.is_some() {
+        plugin_path = Path::new(config.filemaker.ext_path.as_ref().unwrap()).join(plugin_name);
+        remove_dir_all(&plugin_path).ok();
+        create_dir_all(&plugin_path)?;
+    } else {
+        plugin_path = Path::new(out_dir).join(plugin_name);
+        remove_dir_all(&plugin_path).ok();
+        create_dir_all(&plugin_path)?;
+    }
 
-    let bin_path = format!("{}/Contents/MacOS", plugin_path);
+    let bin_path = Path::new(&plugin_path).join("/Contents/MacOS");
     create_dir_all(&bin_path)?;
 
-    let from = format!("{}/lib{}.dylib", out_dir, package_name);
-    let to = format!("{}/{}", bin_path, config.plugin.name);
+    let mut lib_name = String::from("lib");
+    lib_name.push_str(&package_name);
+    let from = Path::new(out_dir).join(lib_name);
+
+    let to = Path::new(&bin_path).join(&config.plugin.name);
     rename(from, to)?;
 
     Ok(())
